Adjust default timeout in PromiseTests
The current value is Inf. However, this results in deadlock if any of
the tests wait on an operation that never completes. This shouldn't
occur in practise, but as an example, a previous version of
scala.concurrent.impl.Future would trigger this for some of the current
tests. This change replaces Inf with a short value, and adds a comment
to explain why this should be safe to do.